+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 7.6.14 (03/18/2021 - )
 * search: update es to 7.12.0
+* action: change error "action log context value too long" to warning
 
 ### 7.6.13 (03/01/2021 - 03/18/2021)